requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.16.1
protobuf==4.25.3
xgboost==2.0.3
ta==0.11.0
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
joblib==1.3.2
